Use entity icon for quick settings tiles by default
 - When creating a new quick setting tile, use the entity's icon by default
 - Move Wear icon based on entity domain code to common
