+++ common/build.gradle.kts
@@ -74,4 +74,7 @@ dependencies {
     implementation("com.squareup.okhttp3:logging-interceptor:4.10.0")
     implementation("com.fasterxml.jackson.module:jackson-module-kotlin:2.13.4")
     implementation("org.altbeacon:android-beacon-library:2.19.4")
+
+    implementation("com.mikepenz:iconics-core:5.4.0")
+    implementation("com.mikepenz:community-material-typeface:7.0.96.0-kotlin@aar")
 }
